refactor(core): 将百分比相关字段和计算统一改为BigDecimal类型处理
- 引入DefValueHelper新增的方法：nvl、calcPercent、normalizeLimit，简化代码并增强复用

+++ mdp-apps/md-console/console-pojo/src/main/java/top/mddata/console/vo/dashboard/DistributionVo.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 分布数据 VO
@@ -23,5 +24,5 @@ public class DistributionVo implements Serializable {
     private Long count;
 
     @Schema(description = "占比")
-    private Double percent;
+    private BigDecimal percent;
 }
